all done

+++ backend/controllers/goalController.js
@@ -2,9 +2,11 @@ const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel.js')
 const { default: mongoose } = require('mongoose')
 
+const User = require('../models/userModel.js')
+
 
 const getGoals =  asyncHandler(async(req,res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id})
 
     res.status(200).json(goals)
 })
@@ -16,7 +18,8 @@ const setGoal = asyncHandler(async(req,res) => {
         throw new Error ('Please add text feild')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -24,9 +27,20 @@ const setGoal = asyncHandler(async(req,res) => {
 
 const updateGoal = asyncHandler(async(req,res) => {
     const goal = await Goal.findById(req.params.id)
-    if (!goal) {
+     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    //Check for User
+     
+    if (!req.user) {
+        res.status(401)
+        throw new Error('User not found')
+      }
+    // Make sure login user matches the goal user
+    if (goal.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true
@@ -41,6 +55,16 @@ const deleteGoal = asyncHandler(async(req,res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+     //Check for User
+     if (!req.user) {
+         res.status(401)
+         throw new Error('User not found')
+       }
+     // Make sure login user matches the goal user
+     if (goal.user.toString() !== req.user.id) {
+         res.status(401)
+         throw new Error('User not authorized')
+     }
     await Goal.findByIdAndDelete(req.params.id)
 
     res.status(200).json({id: req.params.id})
